+++ packages/cf-deploy-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @sap-ux/cf-deploy-config-sub-generator
 
+## 0.2.39
+
+### Patch Changes
+
+-   089b56f: updates extension types and extends prompt options in sub gens
+-   Updated dependencies [089b56f]
+    -   @sap-ux/cf-deploy-config-inquirer@0.4.0
+
+## 0.2.38
+
+### Patch Changes
+
+-   Updated dependencies [ca44076]
+    -   @sap-ux/fiori-generator-shared@0.13.3
+    -   @sap-ux/deploy-config-generator-shared@0.1.22
+    -   @sap-ux/inquirer-common@0.7.27
+    -   @sap-ux/cf-deploy-config-inquirer@0.3.30
+
+## 0.2.37
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.26
+-   @sap-ux/cf-deploy-config-inquirer@0.3.29
+-   @sap-ux/deploy-config-generator-shared@0.1.21
+
 ## 0.2.36
 
 ### Patch Changes